Allows promotion of write knobs to the gizmo.
Allows for  knobs of the internal write node to be promoted to the
interface of the gizmo by way of link knobs that are linked and shown
dynamically at profile load time. The list of knobs to promote are
pulled from the “promote_write_knobs” key that can be defined as part
of the profile’s config definition.