#37034: Resolves a ValueError raised by Nuke's colorspaces.py module.
Recomputing the render paths after restoring cached knob settings caused the exception to be raised. Rather than recompute the paths, we're now removing the the file and proxy knob settings from the cache before applying it. This means we don't need to do the recompute and therefore we no longer see the error.