#39647: Addresses a ValueError raised by Nuke during new node setup. (#21)
In Nuke 10.0, we've seen a number of cases where, as Foundry support explains, the root node for a node graph can be in a state where it's initialized, but doesn't yet have a PythonObject associated with it. This cases some operations from Python to trigger the raising of a ValueError stating that no PythonObject is associated with the node. This exception is not raised in the Python event loop that custom code is running in, which means the user-written code, like tk-nuke-writenode's handler.py, can't catch that exception and handle it gracefully.

The workaround to one of these situations is included here, which is to delay the node creation callback by a short period of time to allow Nuke the chance to complete the initialization of the Python portion of the graph's root node. This is a bummer.